+++ src/main/java/DiffOperator.java
@@ -1,7 +1,7 @@
 
 public class DiffOperator implements Operator {
     @Override
-    public int apply(Operands operands) {
+    public double apply(Operands operands) {
         return operands.var1-operands.var2;
     }
 }
