+++ src/main/java/Operands.java
@@ -1,11 +1,5 @@
-/**
- * This is a concrete class, while in real life, operands can have a single value (e.g. sqrt(2))
- * What can be done to fix it ?
- *
- * I don't know...
- *
- */
+
 public class Operands {
-    public Integer var1;
-    public Integer var2;
+    public Double var1;
+    public Double var2;
 }
